+++ lib/exception.ex
@@ -13,3 +13,11 @@ defmodule Cldr.Calendar.Formatter.InvalidDateError do
     %__MODULE__{message: message}
   end
 end
+
+defmodule Cldr.Calendar.Formatter.InvalidOption do
+  defexception [:message]
+
+  def exception(message) do
+    %__MODULE__{message: message}
+  end
+end