+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 **Note that `ex_cldr_calendars_format` version 1.0.0 and later are supported on Elixir 1.12 and later only.**
 
+## Cldr Calendars Format v1.0.1
+
+This is the changelog for Cldr Calendars Format v1.0.1 released on December 16th, 2025.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_calendars_html/tags)
+
+### Bug Fixes
+
+* Implements the `:day_names` option for `Cldr.Calendar.Formatter.Options` to align with the documentation. Thanks to @allenwyma for the issue. Closes #8.
+
 ## Cldr Calendars Format v1.0.0
 
 This is the changelog for Cldr Calendars Format v1.0.0 released on March 19th, 2025.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_calendars_html/tags)